\docType{data}
\name{all_treebase}
\alias{all_treebase}
\title{treebase.rda}
\description{
  Contains a cache of all phylogenies search_treebase()
  function was able to pull down when run on 2011-12-12.
}
\details{
  recreate with: \code{ all_treebase <-
  search_treebase("Consensus", "type.tree") }
}
\keyword{data}
\keyword{datasets}

