\docType{data}
\name{all_trees}
\alias{all_trees}
\title{branchlengths.rda}
\description{
  All treebase trees with branch length data, created
  2011-12-12
}
\details{
  recreate with code: \code{ all_trees <-
  search_treebase("Consensus", "type.tree",
  branch_lengths=TRUE) }
}
\keyword{data}
\keyword{datasets}

