% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_metadata.R
\name{clean_data}
\alias{clean_data}
\title{clean the fish.base data into pure ASCII}
\usage{
clean_data(metadata)
}
\arguments{
\item{metadata}{list item with fishbase data}
}
\value{
the item scrubbed of non-ASCII characters
}
\description{
clean the fish.base data into pure ASCII
}
\keyword{internal}
