\name{drop_nontrees}
\alias{drop_nontrees}
\title{drop errors from the search}
\usage{
  drop_nontrees(tr)
}
\arguments{
  \item{tr}{a list of phylogenetic trees returned by
  search_treebase}
}
\value{
  the list of phylogenetic trees returned successfully
}
\description{
  drop errors from the search
}
\details{
  primarily for the internal use of search_treebase, but
  may be useful
}

