% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treebase.R
\name{have_branchlength}
\alias{have_branchlength}
\title{Simple function to identify which trees have branch lengths}
\usage{
have_branchlength(trees)
}
\arguments{
\item{trees}{a list of phylogenetic trees (ape/phylo format)}
}
\value{
logical string indicating which have branch length data
}
\description{
Simple function to identify which trees have branch lengths
}
