% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/download_metadata.R
\name{metadata_from_oai}
\alias{metadata_from_oai}
\title{Internal function for OAI-MPH interface to the Dryad database}
\usage{
metadata_from_oai(query, curl = curl)
}
\arguments{
\item{query}{a properly formed url query to dryad}

\item{curl}{if calling in series many times, call getCurlHandle() first and
then pass the return value in here. Avoids repeated handshakes with server.}
}
\description{
Internal function for OAI-MPH interface to the Dryad database
}
\seealso{
\code{\link{dryad_metadata}}
}
\keyword{internal}

