% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_metadata.R
\name{show_metadata}
\alias{show_metadata}
\title{Get the metadata associated with the study in which the phylogeny
 was published.}
\usage{
show_metadata(study.id, curl = getCurlHandle())
}
\arguments{
\item{study.id}{The treebase study id (numbers only, specify in quotes)}

\item{curl}{if calling in series many times, call getCurlHandle()
 first and then pass the return value in here.  avoids repeated
handshakes with server.}
}
\description{
Get the metadata associated with the study in which the phylogeny
 was published.
}
\details{
if the tree is imported with search_treebase, 
then this is in tree$S.id
}
\keyword{internal}
\keyword{utilities}

