\docType{data}
\name{treebase}
\alias{treebase}
\title{treebase.rda}
\description{
  Contains a cache of all phylogenies
  \code{cache_treebase()} function was able to pull down
  when run on 2012-05-14.

  Contains a cache of all publication metadata the
  search_metadata() to pull down when run on 2012-05-12.
}
\details{
  recreate with: \code{ cache_treebase() }

  recreate with: \code{ search_metadata() }
}
\keyword{data}
\keyword{datasets}

