\docType{data}
\name{all_treebase}
\alias{all_treebase}
\title{treebase.rda}
\description{
  Contains a cache of all phylogenies search_treebase()
  function was able to pull down when run on 2012-02-23 to
  2012-02-24.
}
\details{
  recreate with: \code{ cache_treebase() }
}
\keyword{data}
\keyword{datasets}

