% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{treebase}
\alias{treebase}
\title{treebase.rda}
\description{
Contains a cache of all phylogenies \code{cache_treebase()} function was able
to pull down when run  on 2012-05-14.
}
\details{
recreate with: \code{
 cache_treebase() 
}
}
\keyword{data}

